+++ src/app/loadout-builder/generated-sets/CompareDrawer.tsx
@@ -18,7 +18,7 @@ import _ from 'lodash';
 import React, { useMemo, useState } from 'react';
 import { connect } from 'react-redux';
 import { getItemsFromLoadoutItems } from '../../loadout-drawer/loadout-utils';
-import { assignModsToArmorSet } from '../mod-utils';
+import { getAssignedAndUnassignedMods, getModAssignments } from '../mod-assignments';
 import { getTotalModStatChanges } from '../process/mappers';
 import { ArmorSet, ArmorStats, LockableBucketHashes } from '../types';
 import { getPower, upgradeSpendTierToMaxEnergy } from '../utils';
@@ -143,18 +143,17 @@ function CompareDrawer({
     loadoutStats[statHash] += lockedModStats[statHash];
   }
 
-  const [assignedMods] = assignModsToArmorSet(
-    defs,
-    set.armor.map((items) => items[0]),
+  const loSetAssignedMods = getModAssignments(
+    setItems,
     lockedMods,
+    defs,
     upgradeSpendTier,
     lockItemEnergyType
   );
-
-  const [loadoutAssignedMods, loadoutUnassignedMods] = assignModsToArmorSet(
-    defs,
+  const [loadoutAssignedMods, loadoutUnassignedMods] = getAssignedAndUnassignedMods(
     loadoutItems,
     lockedMods,
+    defs,
     upgradeSpendTier,
     lockItemEnergyType
   );
@@ -233,7 +232,7 @@ function CompareDrawer({
             {setItems.map((item) => (
               <div key={item.bucket.hash} className={styles.item}>
                 <ConnectedInventoryItem item={item} />
-                <Sockets item={item} lockedMods={assignedMods[item.id]} />
+                <Sockets item={item} lockedMods={loSetAssignedMods.get(item.id)} />
               </div>
             ))}
           </div>
@@ -273,7 +272,7 @@ function CompareDrawer({
                     style={{ gridColumn: LockableBucketHashes.indexOf(item.bucket.hash) + 1 }}
                   >
                     <ConnectedInventoryItem item={item} />
-                    <Sockets item={item} lockedMods={loadoutAssignedMods[item.id]} />
+                    <Sockets item={item} lockedMods={loadoutAssignedMods.get(item.id)} />
                   </div>
                 ))}
               </div>